% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{dt_wrap}
\alias{dt_wrap}
\title{Wrapper for dt with a non-zero mena and non-1 standard deviation.}
\usage{
dt_wrap(x, df, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric. Where to evaluate the density function.}

\item{df}{A positive numeric. The degrees of freedom.}

\item{mean}{A numeric. The mean of the t. Defaults to 0.}

\item{sd}{A positive numeric. The standard deviation of the
t. Defaults to 1.}
}
\description{
Wrapper for dt with a non-zero mena and non-1 standard deviation.
}

