% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{fn_l}
\alias{fn_l}
\title{Function to minimize wrt the loadings.}
\usage{
fn_l(L, F, Y, beta)
}
\arguments{
\item{L}{A matrix of numerics. n by k. The current loadings.}

\item{F}{A matrix of numerics. k by p. The factors.}

\item{Y}{A matrix of numerics. n by p. The data.}

\item{beta}{A positive numeric. The rate of the variances.}
}
\description{
Function to minimize wrt the loadings.
}

