% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succotash.R
\name{fun_scale}
\alias{fun_scale}
\title{Objective function in intermediate step of EM algorithm when
updating scale.}
\usage{
fun_scale(scale_val, T, resid_vec, tau_seq, sig_diag)
}
\arguments{
\item{scale_val}{A positive numeric. The multiplicative factor to
sig_diag.}

\item{T}{a matrix of numerics. K by p}

\item{resid_vec}{A p-vector of numerics. The current residuals.}

\item{tau_seq}{An K-vector of positive numerics. The mixing
standard deviations, not variances.}

\item{sig_diag}{A p-vector of positive numerics. The estimated
variances, not standard deviations.}
}
\description{
Objective function in intermediate step of EM algorithm when
updating scale.
}

