% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{fun_unif_scale}
\alias{fun_unif_scale}
\title{Criterion function for the scale parameter as an intermediate step
to the EM algorithm when using uniform mixtures.}
\usage{
fun_unif_scale(scale_val, Y, az, sig_diag, Tkj, a_seq, b_seq,
  llike_type = "normal")
}
\arguments{
\item{Y}{A p by 1 matrix of numerics. The data.}

\item{az}{A p by 1 matrix. This is alpha %*% Z}

\item{sig_diag}{A p-vector of numerics. The estimated variances.}

\item{Tkj}{An m by p matrix of numerics. The "T-values".}

\item{a_seq}{A vector of numerics. The non-zero left ends of the
uniforms.}

\item{b_seq}{A vector of numerics. The non-zero right ends of the
uniforms.}

\item{like_type}{A character. Available options are "normal" and "t".}
}
\description{
Criterion function for the scale parameter as an intermediate step
to the EM algorithm when using uniform mixtures.
}

