% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{get_pdiffs}
\alias{get_pdiffs}
\title{Calculate the difference in cdfs.}
\usage{
get_pdiffs(resid_vec, sig_diag, a_seq, b_seq, like_type = "normal")
}
\arguments{
\item{resid_vec}{A vector of numerics such that should be distributed mean zero.}

\item{sig_diag}{A p-vector of numerics. The estimated variances.}

\item{a_seq}{A vector of numerics. The non-zero left ends of the
uniforms.}

\item{b_seq}{A vector of numerics. The non-zero right ends of the
uniforms.}

\item{like_type}{A character. Available options are "normal" and "t".}
}
\description{
Calculate the difference in cdfs.
}

