% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{gr_ab}
\alias{gr_ab}
\title{Gradient of function to optimize wrt alpha and beta.}
\usage{
gr_ab(ab, Y, L, F)
}
\arguments{
\item{ab}{A 2 vector of positive numerics. \code{ab[1]} is the current value
of alpha, \code{ab[2]} is the current value of beta.}

\item{Y}{A matrix of numerics. n by p. The data.}

\item{L}{A matrix of numerics. n by k. The current loadings.}

\item{F}{A matrix of numerics. k by p. The factors.}
}
\description{
Gradient of function to optimize wrt alpha and beta.
}

