% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{llike_unif_simp}
\alias{llike_unif_simp}
\title{Me being as careful as possible when calculating the succotash log-likelihood.}
\usage{
llike_unif_simp(Y, Z, pi_vals, alpha, sig_diag, left_seq, right_seq,
  scale_val = 1)
}
\arguments{
\item{Y}{a p by 1 matrix of numerics.}

\item{Z}{a k by 1 matrix of numerics.}

\item{pi_vals}{A vector of non-negative numerics that sum to
one. The mixing proportions.}

\item{alpha}{A p by k matrix of numerics.}

\item{sig_diag}{A p-vector of numerics.}

\item{left_seq}{The left endpoints of the uniforms.}

\item{right_seq}{The right endpoints of the uniforms}

\item{scale_val}{A positivie numeric.}
}
\description{
Me being as careful as possible when calculating the succotash log-likelihood.
}

