% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{minus_trigamma}
\alias{minus_trigamma}
\title{Trigamma minus a constant.}
\usage{
minus_trigamma(x, emp_var)
}
\arguments{
\item{x}{A positive numeric.}

\item{emp_var}{A numeric.}
}
\description{
Trigamma minus a constant.
}

