% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{mod_fa}
\alias{mod_fa}
\title{Moderated factor analysis. Optimize over loadings, factors, and
hyperparameters of variances.}
\usage{
mod_fa(Y, k, tol = 10^-6, itermax = 100)
}
\arguments{
\item{Y}{A matrix of numerics. n by p.}

\item{k}{A positive integer. The rank of the mean.}

\item{tol}{A positive numeric. The stopping criterion.}

\item{itermax}{A positive integer. The maximum number of iterations
to run through in the optimization.}
}
\description{
Moderated factor analysis. Optimize over loadings, factors, and
hyperparameters of variances.
}

