% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_coordinate.R
\name{normal_llike_grad}
\alias{normal_llike_grad}
\title{Gradient of the log-likelihood wrt Z.}
\usage{
normal_llike_grad(Z, Y, alpha, sig_diag, tau_seq, scale_val, pi_vals)
}
\description{
Gradient of the log-likelihood wrt Z.
}

