% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_coordinate.R
\name{normal_llike_simp}
\alias{normal_llike_simp}
\title{Me being super careful in calculating the normal mixtures likelihood.}
\usage{
normal_llike_simp(Z, Y, alpha, sig_diag, tau_seq, scale_val, pi_vals)
}
\description{
Me being super careful in calculating the normal mixtures likelihood.
}

