% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_coordinate.R
\name{normal_only_z}
\alias{normal_only_z}
\title{Wrapper for \code{\link{succotash_llike}}, but useful when calling
optim.}
\usage{
normal_only_z(Z, pi_vals, scale_val, lambda, alpha, Y, tau_seq, sig_diag)
}
\arguments{
\item{Z}{A k by 1 matrix of numerics.}

\item{pi_vals}{An M vector of numerics that sum to one.}

\item{scale_val}{A positive numeric. The variance inflation
parameter.}
}
\description{
Wrapper for \code{\link{succotash_llike}}, but useful when calling
optim.
}

