% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{only_scale}
\alias{only_scale}
\title{Wrapper for \code{\link{succotash_llike_unif}} but useful for optim
where only update scale_val.}
\usage{
only_scale(scale_val, pi_Z_minus_scale, lambda, alpha, Y, a_seq, b_seq,
  sig_diag)
}
\arguments{
\item{scale_val}{A positive numeric. The variance inflation
parameter.}

\item{pi_Z_minus_scale}{A vector of numerics. The first M are the
pi-vals, the last k are the Z-vals.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{sig_diag}{A vector of length \code{p} containing the variances
of the observations.}
}
\description{
Wrapper for \code{\link{succotash_llike_unif}} but useful for optim
where only update scale_val.
}

