% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{only_Z}
\alias{only_Z}
\title{Wrapper for \code{\link{succotash_llike_unif}} but useful for optim where only update Z.}
\usage{
only_Z(Z, pi_vals, scale_val, lambda, alpha, Y, a_seq, b_seq, sig_diag)
}
\arguments{
\item{Z}{A k by 1 matrix of numerics. The confounders.}

\item{pi_vals}{A M vector of numerics. The mixing probs.}

\item{scale_val}{A positive numeric. The variance inflation parameter.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{sig_diag}{A vector of length \code{p} containing the variances
of the observations.}
}
\description{
Wrapper for \code{\link{succotash_llike_unif}} but useful for optim where only update Z.
}

