% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{pca_naive}
\alias{pca_naive}
\title{Basic PCA}
\usage{
pca_naive(Y, r)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}
}
\description{
Most if not all of code is from package \code{cate}. This is mostly
so people don't have to install sva and leapp if they want to use
it.
}

