% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{pca_shrinkvar}
\alias{pca_shrinkvar}
\title{Use PCA to estimate confounders, then shrink variances using Smyth's method
implemented in limma.}
\usage{
pca_shrinkvar(Y, k, df = "rank_based")
}
\arguments{
\item{Y}{A matrix of numerics. n by p.}

\item{k}{A positive integer. The known rank of the mean.}

\item{df}{A string. Should we subtract off the rank (\code{"rank_based"}) or
just 1 (\code{"minus_one"})?}
}
\description{
Factors are estimated using PCA. Variances are shrunk using an empirical
Bayesian approach assuming a hierarchical model as described in Smyth (2004).
}
\references{
Smyth, G. K. (2004).
  \href{http://www.statsci.org/smyth/pubs/ebayes.pdf}{Linear models and
  empirical Bayes methods for assessing differential expression in microarray
  experiments}. Statistical Applications in Genetics and Molecular Biology
  Volume 3, Issue 1, Article 3.
}

