% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succotash.R
\name{succotash_fixed}
\alias{succotash_fixed}
\title{A fixed-point iteration of the EM algorithm.}
\usage{
succotash_fixed(pi_Z, lambda, alpha, Y, tau_seq, sig_diag,
  plot_new_ests = FALSE, var_scale = TRUE, pen = 0)
}
\arguments{
\item{pi_Z}{A vector. The first \code{M} values are the current
values of \eqn{\pi}. The last \code{k} values are the current
values of \eqn{Z}.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions. If
\code{NULL} then refer to \code{lambda_type}.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k}
and are the coefficients to the confounding variables.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{tau_seq}{A vector of length \code{M} containing the standard
deviations (not variances) of the mixing distributions.}

\item{sig_diag}{A vector of length \code{p} containing the
variances of the observations.}

\item{plot_new_ests}{A logical. Should we plot the new estimates of
pi?}

\item{var_scale}{A logical. Should we update the scaling on the
variances (\code{TRUE}) or not (\code{FALSE}).}

\item{pen}{Defunct. A numeric (usually positive). The correction
term for the scaling factor. This doesn't work too well so may
be removed at any time.}
}
\value{
\code{pi_new} A vector of length \code{M}. The update for the mixing
  components.

  \code{Z_new} A vector of length \code{k}. The update for the confounder
  covariates.
}
\description{
This is a fixed-point iteration for the SUCCOTASH EM algorithm. This updates
the estimte of the prior and the estimate of the hidden covariates.
}

