% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succotash.R
\name{succotash_summaries}
\alias{succotash_summaries}
\title{Provides posterior summaries in the SUCCOTASH model.}
\usage{
succotash_summaries(Y, Z, pi_vals, alpha, sig_diag, tau_seq)
}
\arguments{
\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{Z}{A matrix of dimension \code{k} by \code{1}. The
(estimated) confounding covariates.}

\item{pi_vals}{A vector of length \code{M}. The (estimated) mixing
proportions.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k}
and are the coefficients to the confounding variables.}

\item{sig_diag}{A vector of length \code{p} containing the
variances of the observations.}

\item{tau_seq}{A vector of length \code{M} containing the standard
deviations (not variances) of the mixing distributions.}
}
\value{
\code{lfdr} (local false discovery rate) A vector of length
    \code{p}.  The posterior probability that \eqn{\beta_j = 0}.

  \code{lfsr} (local false sign rate) A vector of length
  \code{p}. The posterior probability of making a sign error.

  \code{betahat} A vector of length \code{p}. The posterior
  estimates of \eqn{\beta}.
}
\description{
\code{succotash_summaries} will return useful posterior summaries
used in estimation and testing/FDR control.
}
\details{
The posterior distribution is just a mixture of normals. This
function will the posterior means as a point estimate. It will also
return local false sign and discovery rates. These are useful for
controlling for multiple testing.
}

