% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{succotash_unif_fixed}
\alias{succotash_unif_fixed}
\title{A fixed point iteration in the mixture of uniforms EM.}
\usage{
succotash_unif_fixed(pi_Z, lambda, alpha, Y, a_seq, b_seq, sig_diag,
  print_ziter = FALSE, newt_itermax = 4, tol = 10^-4, var_scale = TRUE)
}
\arguments{
\item{pi_Z}{A vector. The first \code{M} values are the current
values of \eqn{\pi}. The last \code{k} values are the current
values of \eqn{Z}.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{sig_diag}{A vector of length \code{p} containing the
variances of the observations.}

\item{print_ziter}{A logical. Should we we print each iteration of
the Z optimization?}

\item{newt_itermax}{A positive integer. The maximum number of Newton
steps to perform in updating \eqn{Z}.}

\item{tol}{A positive numeric. The stopping criterion for Newton's
method in updating Z.}

\item{var_scale}{A logical. Should we update the scaling on the
variances (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
\code{pi_new} A vector of length \code{M}. The update for
    the mixing components.

 \code{Z_new} A vector of length \code{k}. The update for the
 confounder covariates.
}
\description{
This is a fixed-point iteration for the SUCCOTASH EM algorithm. This
updates the estimate of the prior and the estimate of the hidden
covariates.
}
\seealso{
\code{\link{uniform_succ_given_alpha}}
    \code{\link{succotash_llike_unif}}
}

