% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_unif_succ.R
\name{t_unif_em}
\alias{t_unif_em}
\title{EM algorithm for uniform mixtures and t-likelihood}
\usage{
t_unif_em(a_seq, b_seq, Y, alpha, sig_diag, nu, pi_init, Z_init, pi_init_type,
  lambda, print_progress, print_ziter, em_z_start_sd, true_Z = NULL,
  em_tol = 10^-6, em_itermax = 200)
}
\arguments{
\item{a_seq}{A vector of negative numerics in increasing order. The
negative end points in an [a, 0] grid.}

\item{b_seq}{A vector of positive numerics in increasing order. The
positive end points in a [0, b] grid.}

\item{Y}{A p by 1 matrix of numerics. The data.}

\item{alpha}{A p by k matrix of numerics. The confounder
coefficients.}

\item{sig_diag}{A vector of the variances of \code{Y}.}

\item{nu}{A positive numeric. The degrees of freedom of the
t-likelihood.}

\item{pi_init}{A vector of non-negative numerics that sum of 1 of
length \code{length(a) + length(b) + 1}. The initial values of
the mixture probs.}

\item{Z_init}{A vector of length k of numerics. Starting values of
Z.}

\item{pi_init_type}{Either "random", "uniform", or "zero_conc". How
should we choose the initial mixture probabilities if pi_init
is NULL? "random" will draw draw pi uniformly from the
simplex. "uniform" will give each value equal mass. "zero_conc"
will give more mass to 0 than any other probability.}

\item{lambda}{A vector of numerics greater than or equal to 1, of
length \code{length(a) + length(b) + 1}.}

\item{print_progress}{A logical. Should we plot the progress?}

\item{print_ziter}{A logical. Should we print the progress of the
Newton iterations for updating Z?}

\item{em_z_start_sd}{A positive numeric. Z is initialized by iid
normals with this standard deviation and mean 0.}

\item{true_Z}{The true Z values. Used for testing.}

\item{em_tol}{A positive numeric. The stopping criterion for the EM
algorithm.}

\item{em_itermax}{A positive integer. The maximum number of
iterations to perform on the em step.}
}
\description{
EM algorithm for uniform mixtures and t-likelihood
}

