% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_unif_succ.R
\name{tfun}
\alias{tfun}
\title{Function to optimize over Z in EM step.}
\usage{
tfun(Z_old, sig_diag, Y, alpha, Tkj, a_seq, b_seq, nu)
}
\arguments{
\item{Z_old}{A vector of numerics. The current value of Z.}

\item{sig_diag}{A vector of length \code{p} containing the variances
of the observations.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{Tkj}{A matrix of numerics. The weights from pi_new.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{nu}{A positive numeric. The degrees of freedom of the
t-distribution.}
}
\description{
Function to optimize over Z in EM step.
}

