% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_unif_succ.R
\name{tupdate_pi}
\alias{tupdate_pi}
\title{EM step to update pi}
\usage{
tupdate_pi(Z_old, pi_old, sig_diag, Y, alpha, a_seq, b_seq, nu, lambda)
}
\arguments{
\item{Z_old}{A vector of numerics. The current value of Z.}

\item{pi_old}{A vector of numerics that sums to one. The current value of pi.}

\item{sig_diag}{A vector of length \code{p} containing the variances
of the observations.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{nu}{A positive numeric. The degrees of freedom of the
t-distribution.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}
}
\description{
EM step to update pi
}

