% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{unif_grad_llike}
\alias{unif_grad_llike}
\title{My first attempt at calculating the gradient of the likelihood
function using uniform mixtures wrt Z.}
\usage{
unif_grad_llike(pi_Z, lambda, alpha, Y, a_seq, b_seq, sig_diag,
  var_scale = TRUE)
}
\arguments{
\item{pi_Z}{A vector. The first \code{M} values are the current
values of \eqn{\pi}. The last \code{k} values are the current
values of \eqn{Z}.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{sig_diag}{A vector of length \code{p} containing the variances
of the observations.}

\item{var_scale}{A logical. Should we update the scaling on the
variances (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
This doesn't seem to work too well.
}

