% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{unif_update_pi}
\alias{unif_update_pi}
\title{E step in EM algorithm.}
\usage{
unif_update_pi(pi_old, Y, sig_diag, alpha, a_seq, b_seq, Z_old, lambda)
}
\arguments{
\item{pi_old}{A vector of numerics. The old pi values.}

\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{sig_diag}{A vector of length \code{p} containing the
variances of the observations.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{Z_old}{A vector of numerics. The old confounders.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}
}
\description{
E step in EM algorithm.
}

