% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{unif_update_z}
\alias{unif_update_z}
\title{Runs a few newton steps to update Z given scale_val.}
\usage{
unif_update_z(Y, sig_diag, alpha, lambda, a_seq, b_seq, pi_new, Tkj, Z_old,
  llike_old, tol = 10^-3, newt_itermax = 10, print_ziter = FALSE,
  scale_val = 1)
}
\arguments{
\item{Y}{A matrix of dimension \code{p} by \code{1}. These are the
observed regression coefficients of the observed variables.}

\item{sig_diag}{A vector of length \code{p} containing the
variances of the observations.}

\item{alpha}{A matrix. This is of dimension \code{p} by \code{k} and
are the coefficients to the confounding variables.}

\item{lambda}{A vector. This is a length \code{M} vector with the
regularization parameters for the mixing proportions.}

\item{a_seq}{A vector of negative numerics containing the left
endpoints of the mixing uniforms.}

\item{b_seq}{A vector of positiv numerics containing the right
endpoints of the mixing uniforms.}

\item{pi_new}{A vector of numerics. The current estimates of of the
mixing proportions.}

\item{Tkj}{A matrix of numerics. The T matrix.}

\item{Z_old}{A vector of numerics. The old confounders.}

\item{tol}{A positive numeric. The stopping criterion for Newton's
method in updating Z.}

\item{newt_itermax}{A positive integer. The maximum number of Newton
steps to perform in updating \eqn{Z}.}

\item{print_ziter}{A logical. Should we we print each iteration of
the Z optimization?}
}
\description{
Runs a few newton steps to update Z given scale_val.
}

