% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{uniform_succ_em}
\alias{uniform_succ_em}
\title{EM algorithm for second step of SUCCOTASH}
\usage{
uniform_succ_em(pi_init, Z_init, a_seq, b_seq, lambda, alpha, Y, sig_diag,
  print_ziter, print_progress, em_z_start_sd, em_itermax = 200,
  em_tol = 10^-6, pi_init_type = "random", true_Z = NULL)
}
\arguments{
\item{pi_init}{A vector of non-negative numerics that sum of 1 of
length \code{length(a) + length(b) + 1}. The initial values of
the mixture probs.}

\item{Z_init}{A vector of length k of numerics. Starting values of
Z.}

\item{a_seq}{A vector of negative numerics in increasing order. The
negative end points in an [a, 0] grid.}

\item{b_seq}{A vector of positive numerics in increasing order. The
positive end points in a [0, b] grid.}

\item{lambda}{A vector of numerics greater than or equal to 1, of
length \code{length(a) + length(b) + 1}.}

\item{alpha}{A p by k matrix of numerics. The confounder
coefficients.}

\item{Y}{A p by 1 matrix of numerics. The data.}

\item{sig_diag}{A vector of the variances of \code{Y}.}

\item{print_ziter}{A logical. Should we print the progress of the
Newton iterations for updating Z?}

\item{print_progress}{A logical. Should we plot the progress?}

\item{em_z_start_sd}{A positive numeric. Z is initialized by iid
normals with this standard deviation and mean 0.}

\item{em_itermax}{A positive integer. The maximum number of
iterations to perform on the em step.}

\item{em_tol}{A positive numeric. The stopping criterion for the EM
algorithm.}

\item{pi_init_type}{Either "random", "uniform", or "zero_conc". How
should we choose the initial mixture probabilities if pi_init
is NULL? "random" will draw draw pi uniformly from the
simplex. "uniform" will give each value equal mass. "zero_conc"
will give more mass to 0 than any other probability.}

\item{true_Z}{The true Z values. Used for testing.}
}
\description{
EM algorithm for second step of SUCCOTASH
}

