% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mle.R
\name{update_A}
\alias{update_A}
\title{Update low rank mean in regularized maximum likelihood factor analysis.}
\usage{
update_A(Y, sig_diag, k)
}
\arguments{
\item{Y}{A matrix. This is the \code{n} by \code{p} data matrix.}

\item{sig_diag}{A vector of length \code{p}. These are the estimates of the
variances of the columns of \code{Y}.}

\item{k}{An integer. The rank of the mean matrix.}
}
\value{
\code{A_new} The update of the mean matrix.
}
\description{
This is the update of the low rank mean in the regularized factor analysis.
The low rank mean is not regularized, so this is the exact same update as in
\href{http://projecteuclid.org/euclid.aoas/1356629055}{Sun et al (2012)}.
}
\details{
The update is just a truncated SVD of a scaled data matrix, where the scaling
depends on the current estimates of the variances of the columns.
}

