% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{update_ab}
\alias{update_ab}
\title{Updates the hyperparameters for the gamma prior.}
\usage{
update_ab(alpha, beta, Y, L, F)
}
\arguments{
\item{alpha}{A positive numeric. The shape parameter of the variances.}

\item{beta}{A positive numeric. The rate of the variances.}

\item{Y}{A matrix of numerics. n by p. The data.}

\item{L}{A matrix of numerics. n by k. The current loadings.}

\item{F}{A matrix of numerics. k by p. The factors.}
}
\description{
Updates the hyperparameters for the gamma prior.
}

