% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{update_f}
\alias{update_f}
\title{Update the factors given the loadings and the data matrix.}
\usage{
update_f(L, Y)
}
\arguments{
\item{L}{A matrix of numerics. n by k. The current loadings.}

\item{Y}{A matrix of numerics. The data matrix. n by p.}
}
\description{
Update the factors given the loadings and the data matrix.
}

