% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{update_l}
\alias{update_l}
\title{Update the loadings given the factors, the data matrix, and the
rate of the variances.}
\usage{
update_l(L, F, Y, beta)
}
\arguments{
\item{L}{A matrix of numerics. n by k. The current loadings.}

\item{F}{A matrix of numerics. k by p. The factors.}

\item{Y}{A matrix of numerics. n by p. The data.}

\item{beta}{A positive numeric. The rate of the variances.}
}
\description{
Update the loadings given the factors, the data matrix, and the
rate of the variances.
}

