% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{fit_succotash_unif_coord}
\alias{fit_succotash_unif_coord}
\title{Fixed point for uniform succotash when doing coordinate descent.}
\usage{
fit_succotash_unif_coord(pi_Z, lambda, alpha, Y, a_seq, b_seq, sig_diag,
  print_ziter = FALSE, newt_itermax = 100, tol = 10^-4,
  var_scale = TRUE)
}
\description{
Fixed point for uniform succotash when doing coordinate descent.
}

