% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_coordinate.R
\name{only_scale}
\alias{only_scale}
\title{Wrapper for \code{\link{succotash_llike_unif}} but useful for optim where only update scale_val.}
\usage{
only_scale(scale_val, pi_Z_minus_scale, lambda, alpha, Y, a_seq, b_seq,
  sig_diag)
}
\description{
Wrapper for \code{\link{succotash_llike_unif}} but useful for optim where only update scale_val.
}

