% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fa.R
\name{pca_naive}
\alias{pca_naive}
\title{Basic PCA.}
\usage{
pca_naive(Y, r)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}
}
\description{
Most of this code is from the package \code{cate}. I corrected some
problems. Specifically, I allow \code{r = 0} and I included a few
needed \code{drop = FALSE} terms. I also divide by \code{nrow(Y) -
r} rather than by \code{nrow(Y)}.
}
\author{
David Gerard
}

