% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_succ.R
\name{uniform_succ_em}
\alias{uniform_succ_em}
\title{EM algorithm for second step of SUCCOTASH}
\usage{
uniform_succ_em(Y, alpha, sig_diag, a_seq, b_seq, pi_init = NULL,
  Z_init = NULL, lambda = NULL, print_ziter = FALSE,
  print_progress = FALSE, em_z_start_sd = 1, em_itermax = 200,
  em_tol = 10^-3, pi_init_type = c("random", "uniform", "zero_conc"),
  true_Z = NULL, var_scale = TRUE, optmethod = c("coord", "em"))
}
\arguments{
\item{Y}{A p by 1 matrix of numerics. The data.}

\item{alpha}{A p by k matrix of numerics. The confounder
coefficients.}

\item{sig_diag}{A vector of the variances of \code{Y}.}

\item{a_seq}{A vector of negative numerics in increasing order. The
negative end points in an [a, 0] grid.}

\item{b_seq}{A vector of positive numerics in increasing order. The
positive end points in a [0, b] grid.}

\item{pi_init}{A vector of non-negative numerics that sum of 1 of
length \code{length(a) + length(b) + 1}. The initial values of
the mixture probs.}

\item{Z_init}{A vector of length k of numerics. Starting values of
Z.}

\item{lambda}{A vector of numerics greater than or equal to 1, of
length \code{length(a) + length(b) + 1}.}

\item{print_ziter}{A logical. Should we print the progress of the
Newton iterations for updating Z?}

\item{print_progress}{A logical. Should we plot the progress?}

\item{em_z_start_sd}{A positive numeric. Z is initialized by iid
normals with this standard deviation and mean 0.}

\item{em_itermax}{A positive integer. The maximum number of
iterations to perform on the em step.}

\item{em_tol}{A positive numeric. The stopping criterion for the EM
algorithm.}

\item{pi_init_type}{Either "random", "uniform", or "zero_conc". How
should we choose the initial mixture probabilities if pi_init
is NULL? "random" will draw draw pi uniformly from the
simplex. "uniform" will give each value equal mass. "zero_conc"
will give more mass to 0 than any other probability.}

\item{true_Z}{The true Z values. Used for testing.}

\item{var_scale}{A logical. Should we update the scaling on the
variances (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
EM algorithm for second step of SUCCOTASH
}

